@@ -1,3 +1,3 @@
 module github.com/bsm/gomega
 
-go 1.14
+go 1.18
